pyproject.toml
[project]
name = "klub"
version = "0.1.0"
description = ""
authors = [
    {name = "aziret"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
